Input validation on the GraphQL side

+++ auth-server-manage/src/main/java/com/unitvectory/auth/server/manage/controller/ClientResolver.java
@@ -23,7 +23,7 @@ import org.springframework.graphql.data.method.annotation.SchemaMapping;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.oauth2.jwt.Jwt;
 import org.springframework.stereotype.Controller;
-
+import com.unitvectory.auth.common.InputPatterns;
 import com.unitvectory.auth.server.manage.dto.AuthorizationType;
 import com.unitvectory.auth.server.manage.dto.ClientManagementCapabilitiesType;
 import com.unitvectory.auth.server.manage.dto.ClientScopeType;
@@ -33,6 +33,7 @@ import com.unitvectory.auth.server.manage.dto.ResponseType;
 import com.unitvectory.auth.server.manage.mapper.RequestJwtMapper;
 import com.unitvectory.auth.server.manage.service.ClientService;
 import com.unitvectory.auth.server.manage.service.ManagementCapabilitiesService;
+import com.unitvectory.auth.util.exception.BadRequestException;
 
 /**
  * The GraphQL Client Resolver
@@ -51,17 +52,51 @@ public class ClientResolver {
 	@MutationMapping
 	public ClientType addClient(@Argument String clientId, @Argument String description,
 			@Argument List<ClientScopeType> availableScopes) {
+
+		// Input Validation
+
+		if (clientId == null || !clientId.matches(InputPatterns.CLIENT_ID)) {
+			throw new BadRequestException("Invalid 'client_id' attribute format.");
+		} else if (description != null && !description.matches(InputPatterns.DESCRIPTION)) {
+			throw new BadRequestException("Invalid 'description' attribute format.");
+		}
+
+		if (availableScopes != null) {
+			for (ClientScopeType scope : availableScopes) {
+				if (scope == null || !scope.getScope().matches(InputPatterns.SCOPE)) {
+					throw new BadRequestException("Invalid 'scope' attribute format.");
+				} else if (scope == null
+						|| !scope.getDescription().matches(InputPatterns.DESCRIPTION)) {
+					throw new BadRequestException("Invalid 'description' attribute format.");
+				}
+			}
+		}
+
 		return this.clientService.addClient(clientId, description, availableScopes);
 	}
 
 	@MutationMapping
 	public ResponseType deleteClient(@Argument String clientId, @AuthenticationPrincipal Jwt jwt) {
+
+		// Input Validation
+
+		if (clientId == null || !clientId.matches(InputPatterns.CLIENT_ID)) {
+			throw new BadRequestException("Invalid 'client_id' attribute format.");
+		}
+
 		return this.clientService.deleteClient(clientId, RequestJwtMapper.INSTANCE.requestJwt(jwt));
 	}
 
 	@MutationMapping
 	public ClientSecretType generateClientSecret1(@Argument String clientId,
 			@AuthenticationPrincipal Jwt jwt) {
+
+		// Input Validation
+
+		if (clientId == null || !clientId.matches(InputPatterns.CLIENT_ID)) {
+			throw new BadRequestException("Invalid 'client_id' attribute format.");
+		}
+
 		return this.clientService.generateClientSecret1(clientId,
 				RequestJwtMapper.INSTANCE.requestJwt(jwt));
 	}
@@ -69,6 +104,13 @@ public class ClientResolver {
 	@MutationMapping
 	public ClientSecretType generateClientSecret2(@Argument String clientId,
 			@AuthenticationPrincipal Jwt jwt) {
+
+		// Input Validation
+
+		if (clientId == null || !clientId.matches(InputPatterns.CLIENT_ID)) {
+			throw new BadRequestException("Invalid 'client_id' attribute format.");
+		}
+
 		return this.clientService.generateClientSecret2(clientId,
 				RequestJwtMapper.INSTANCE.requestJwt(jwt));
 	}
@@ -76,6 +118,13 @@ public class ClientResolver {
 	@MutationMapping
 	public ClientSecretType clearClientSecret1(@Argument String clientId,
 			@AuthenticationPrincipal Jwt jwt) {
+
+		// Input Validation
+
+		if (clientId == null || !clientId.matches(InputPatterns.CLIENT_ID)) {
+			throw new BadRequestException("Invalid 'client_id' attribute format.");
+		}
+
 		return this.clientService.clearClientSecret1(clientId,
 				RequestJwtMapper.INSTANCE.requestJwt(jwt));
 	}
@@ -83,6 +132,13 @@ public class ClientResolver {
 	@MutationMapping
 	public ClientSecretType clearClientSecret2(@Argument String clientId,
 			@AuthenticationPrincipal Jwt jwt) {
+
+		// Input Validation
+
+		if (clientId == null || !clientId.matches(InputPatterns.CLIENT_ID)) {
+			throw new BadRequestException("Invalid 'client_id' attribute format.");
+		}
+
 		return this.clientService.clearClientSecret2(clientId,
 				RequestJwtMapper.INSTANCE.requestJwt(jwt));
 	}
@@ -91,6 +147,21 @@ public class ClientResolver {
 	public ResponseType authorizeJwtBearer(@Argument String clientId, @Argument String jwksUrl,
 			@Argument String iss, @Argument String sub, @Argument String aud,
 			@AuthenticationPrincipal Jwt jwt) {
+
+		// Input Validation
+
+		if (clientId == null || !clientId.matches(InputPatterns.CLIENT_ID)) {
+			throw new BadRequestException("Invalid 'client_id' attribute format.");
+		} else if (jwksUrl == null || !jwksUrl.matches(InputPatterns.EXTERNAL_JWKS_URL)) {
+			throw new BadRequestException("Invalid 'jwks_url' attribute format.");
+		} else if (iss == null || !iss.matches(InputPatterns.EXTERNAL_CLAIM)) {
+			throw new BadRequestException("Invalid 'iss' attribute format.");
+		} else if (sub == null || !sub.matches(InputPatterns.EXTERNAL_CLAIM)) {
+			throw new BadRequestException("Invalid 'sub' attribute format.");
+		} else if (aud == null || !aud.matches(InputPatterns.EXTERNAL_CLAIM)) {
+			throw new BadRequestException("Invalid 'aud' attribute format.");
+		}
+
 		return this.clientService.authorizeJwtBearer(clientId, jwksUrl, iss, sub, aud,
 				RequestJwtMapper.INSTANCE.requestJwt(jwt));
 	}
@@ -98,6 +169,15 @@ public class ClientResolver {
 	@MutationMapping
 	public ResponseType deauthorizeJwtBearer(@Argument String clientId, @Argument String id,
 			@AuthenticationPrincipal Jwt jwt) {
+
+		// Input Validation
+
+		if (clientId == null || !clientId.matches(InputPatterns.CLIENT_ID)) {
+			throw new BadRequestException("Invalid 'client_id' attribute format.");
+		} else if (id == null || !id.matches(InputPatterns.V4UUID)) {
+			throw new BadRequestException("Invalid 'id' attribute format.");
+		}
+
 		return this.clientService.deauthorizeJwtBearer(clientId, id,
 				RequestJwtMapper.INSTANCE.requestJwt(jwt));
 	}
